Implement purgeOnJoin
The joiner will still do the preload, but the entries will be removed from
the data container.

Also includes a fix for ISPN-3217 (entries not written to the cache store
on a backup owner with CACHE_MODE_LOCAL).

A hack was needed to avoid triggering ISPN-3443 (state transfer undoing the
remove operations).